src/requirements.txt
flask
pillow
python-docx